# functions/src/api/api-swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Transcriber API",
        "version": "2.0",
        "description" : "### Authentication\n\n* Setup you application to use [Azure AD JWT Token](https://docs.microsoft.com/en-us/azure/active-directory/develop/id-tokens) or\n* Setup your application to use [Firebase JWT Custom Token](https://firebase.google.com/docs/auth/admin/create-custom-tokens) or\n* Use [Transcriber Cli](https://github.com/nrkno/transcriber-cli) to generate Firebase Custom Token "
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/transcriptId": {
            "post": {
                "summary": "Create a new ID. This ID is required for creating upload URL later.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "operationId": "createTranscriptId",
                "tags": [
                    "transcriptId"
                ],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Expected response to a valid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TranscriptId"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/UnAuthenticatedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/UnAuthorizedError"
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/uploadUrl": {
            "post": {
                "summary": "Create url you can use to upload the media file to Google Storage.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "operationId": "createUploadUrl",
                "tags": [
                    "uploadUrl"
                ],
                "parameters": [
                    {
                        "name": "transcriptId",
                        "in": "query",
                        "required": true,
                        "description": "The transcriptId generated from the /api/transcriptId call.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-Content-Type",
                        "in": "header",
                        "required": true,
                        "description": "Content Type need to match the Content-Type used when you upload the file. Eg: audio/mpeg.",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Expected response to a valid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UploadUrl"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Response as redirect url, when acceapt header != application/json"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnAuthenticatedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/UnAuthorizedError"
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/uploadFile" : {
            "post" : {
                "tags" : [ "uploadFile" ],
                "summary" : "You need to use eg. curl to upload a file using the url you got from /uploadUrl",
                "description" : "### Upload File with Curl \n* exportFILE_PATH=<full_path>\n* export UPLOAD_URL=<from_uploadUrl_above<\n* CONTENT_TYPE=<as_used_in_uploadUrl> eg.audio/mpeg\n\n ```curl -X PUT --data-binary @$FILE_PATH -H \"Content-Type: $CONTENT_TYPE\" $UPLOAD_URL```\n"
            }
        },
        "/transcripts/{transcriptId}": {
            "post": {
                "summary": "Create transcription details. Require that a file has been uploaded using /uploadUrl. When these details are created, transcription will start automatically.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "operationId": "createTranscription",
                "tags": [
                    "transcripts"
                ],
                "parameters": [
                    {
                        "name": "transcriptId",
                        "in": "path",
                        "required": true,
                        "description": "The id of the transcript",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Optional description in *Markdown*",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateTranscription"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "202": {
                        "description": "Follow location header to find transcription status."
                    },
                    "401": {
                        "$ref": "#/components/responses/UnAuthenticatedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/UnAuthorizedError"
                    },
                    "422": {
                        "$ref": "#/components/responses/RequiredParametersMissingError"
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "summary": "Fetch the transcription details.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "operationId": "getTranscriptById",
                "tags": [
                    "transcripts"
                ],
                "parameters": [
                    {
                        "name": "transcriptId",
                        "in": "path",
                        "required": true,
                        "description": "The id of the transcript to fetch",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Expected response to a valid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transcript"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/UnAuthenticatedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/UnAuthorizedError"
                    },
                    "404": {
                        "$ref": "#/components/responses/TranscriptNotFoundError"
                    },
                    "422": {
                        "$ref": "#/components/responses/RequiredParametersMissingError"
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/transcripts/{transcriptId}/export": {
            "get": {
                "summary": "Export transcribed words to Json, Docx or XMP",
                "description": "### Export Format\nSelect Accept header to either\n* application/json\n* application/docx\n* application/xmp",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "operationId": "exportTranscriptById",
                "tags": [
                    "export"
                ],
                "parameters": [
                    {
                        "name": "transcriptId",
                        "in": "path",
                        "required": true,
                        "description": "The id of the transcript to export",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Expected response to a valid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExportedTranscriptJson"
                                }
                            },
                            "application/docx": {
                                "schema": {
                                    "type": "string",
                                    "format": "xml"
                                }
                            },
                            "application/xmp": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/UnAuthenticatedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/UnAuthorizedError"
                    },
                    "404": {
                        "$ref": "#/components/responses/TranscriptNotFoundError"
                    },
                    "422": {
                        "$ref": "#/components/responses/RequiredParametersMissingError"
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "TranscriptId": {
                "required": [
                    "transcriptId"
                ],
                "properties": {
                    "transcriptId": {
                        "type": "string"
                    }
                }
            },
            "CreateTranscription": {
                "required": [
                    "originalMimeType"
                ],
                "properties": {
                    "originalMimeType": {
                        "type": "string",
                        "description": "The Content-Type used when uploading the audio or video file."
                    }
                }
            },
            "Transcript": {
                "required": [
                    "id",
                    "status",
                    "userId"
                ],
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "status": {
                        "type": "object",
                        "properties": {
                            "progress": {
                                "type": "string"
                            }
                        }
                    },
                    "userId": {
                        "type": "string"
                    },
                    "paragraphs": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Paragraph"
                        }
                    },
                    "metadata": {
                        "type": "object",
                        "properties": {
                            "audioDuration": {
                                "type": "integer"
                            },
                            "languageCodes": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            },
                            "originalMimeType": {
                                "type": "string"
                            }
                        }
                    },
                    "playbackGsUrl": {
                        "type": "string",
                        "format": "uri"
                    }
                }
            },
            "Paragraph": {
                "properties": {
                    "startTime": {
                        "type": "integer"
                    },
                    "words": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Word"
                        }
                    }
                }
            },
            "Word": {
                "properties": {
                    "endTime": {
                        "type": "integer"
                    },
                    "confidence": {
                        "type": "number"
                    },
                    "startTime": {
                        "type": "integer"
                    },
                    "text": {
                        "type": "string"
                    }
                }
            },
            "TransctriptList": {
                "type": "array",
                "items": {
                    "$ref": "#/components/schemas/Transcript"
                }
            },
            "ExportedTranscriptJson": {
                "type": "array",
                "items": {
                    "$ref": "#/components/schemas/ExportedTranscriptJsonWords"
                }
            },
            "ExportedTranscriptJsonWords": {
                "properties": {
                    "startTime": {
                        "type": "string"
                    },
                    "words": {
                        "type": "string"
                    }
                }
            },
            "UploadUrl": {
                "required": [
                    "uploadUrl"
                ],
                "properties": {
                    "uploadUrl": {
                        "type": "string"
                    }
                }
            },
            "Error": {
                "required": [
                    "code",
                    "message"
                ],
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
        "responses": {
            "UnAuthenticatedError": {
                "description": "Please add a valid token as header. Authorization: Bearer your_token . <br/>Both Azure AD Access Token and Firebase Custom Token is accepted."
            },
            "UnAuthorizedError": {
                "description": "Please add a valid token as header. Authorization: Bearer your_token . <br/>Both Azure AD Access Token and Firebase Custom Token is accepted."
            },
            "TranscriptNotFoundError": {
                "description": "No Transcript was found."
            },
            "RequiredParametersMissingError": {
                "description": "Required Parameters are missing in your request. Each response will notify the client of what info is missing."
            }
        }
    }
}